change component archive name and move implementation to container
