+++ client/src/containers/rulesListContainer.js
@@ -1,18 +1,84 @@
 import React from 'react'
-import RulesListForm from '../components/rulesListForm'
+import RulesList from '../components/rulesList'
+import Http from '../service/Http'
+import RuleTranslator from '../service/ruleTranslator'
 
 export default class RulesContainer extends React.Component {
   constructor (props) {
     super(props)
     this.state = {
-      errors: {}
+      errors: {},
+      haveTheRules: false,
+      refresh: false,
+      rules: [{
+        'rule': [],
+        'coloquialRule': '',
+        'id': ''
+      }]
     }
   }
 
+  handleChange = (name, value) => {
+    this.setState({
+      [name]: value
+    })
+  }
+
+  getRules = () => {
+    if (this.state.refresh) return
+    let rulesVector = []
+    Http.get('/rules/')
+      .then(response => {
+        // expected: { success: true, rules: rules }
+        if (response.success) {
+          const length = response.rules.length
+          for (let i = 0; i < length; i++) {
+            const aRule = response.rules[i].json
+            const ruleTranslated = RuleTranslator.translateRule(aRule)
+            const id = response.rules[i].id
+            rulesVector.push(
+              {
+                'rule': aRule,
+                'coloquialRule': ruleTranslated,
+                'id': id
+              })
+          }
+          this.handleChange('rules', rulesVector)
+          this.handleChange('refresh', true)
+        } else {
+          alert('Error: 200 status was not received')
+        }
+      })
+      .catch(err => {
+        alert(err)
+      })
+  }
+
+  deleteRule = (aRule) => {
+    Http.delete('/rules/', aRule.id)
+      .then(response => {
+        // expected: { success: true }
+        if (response.success) {
+          this.setState({ 'refresh': true })
+          alert('The deletion succeded')
+        } else {
+          alert('The deletion did not succeded')
+        }
+      })
+      .catch(err => {
+        alert(err)
+      })
+  }
+
   render () {
-    const { errors } = this.state
+    const { errors, rules } = this.state
     return (
-      <RulesListForm errors={errors} onClick={(rules) => this.handleClick(rules)}/>
+      <RulesList 
+        errors={errors}
+        rules={rules}
+        getRules={(rules) => this.getRules(rules)}
+        deleteRule={(aRule => this.deleteRule(aRule))}
+      />
     )
   }
 }
